+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Option<T>.map` function with unit tests and docs
 - `Option<T>.or` function with unit tests and docs
 - `Option<T>and` function with unit tests and docs
+- `Option<T>.andThen` function with unit tests and docs
 
 ## [0.3.0] - 27/05/2020